WIP. Group test runs
When running on a project with many small tests,
I noticed that running parallel increased the total
duration - this turns out to be because the cost of
launching the binary was (significantly) greater than
the cost of most of the tests.

This patch allows the tests to be grouped by suite
such that fewer processes need launching, but
parallelisation is still possible (and now becomes
beneficial).

Test-Information:
Without this change, runtime with gtest-parallel
was increased about 4x compared to without. With
this change it's between halved and quartered
instead, on those machines I tested.